Get the UI aspects basically working
Now the letters are displayed in a circle and you can drag between them,
leaving a trail highlighting the letters selected so far.

This code is shamefully messy - my apologies. The non-prototype version
of this will have an actual design.

Also, I've decided to use Optional but I haven't bothered to fix the
minimum API version, so don't try this on a lower-API device.

TODO:
1. Notify the callback with the highlighted word on release
2. Disallow selecting the same letter twice